+++ src/main/java/com/nisovin/shopkeepers/commands/shopkeepers/CommandTestDamage.java
@@ -1,0 +1,117 @@
+package com.nisovin.shopkeepers.commands.shopkeepers;
+
+import org.bukkit.ChatColor;
+import org.bukkit.entity.LivingEntity;
+import org.bukkit.entity.Player;
+import org.bukkit.scheduler.BukkitRunnable;
+
+import com.nisovin.shopkeepers.SKShopkeepersPlugin;
+import com.nisovin.shopkeepers.api.ShopkeepersPlugin;
+import com.nisovin.shopkeepers.commands.lib.CommandContextView;
+import com.nisovin.shopkeepers.commands.lib.CommandException;
+import com.nisovin.shopkeepers.commands.lib.CommandInput;
+import com.nisovin.shopkeepers.commands.lib.PlayerCommand;
+import com.nisovin.shopkeepers.commands.lib.arguments.BoundedDoubleArgument;
+import com.nisovin.shopkeepers.commands.lib.arguments.DefaultValueFallback;
+import com.nisovin.shopkeepers.commands.lib.arguments.PositiveIntegerArgument;
+import com.nisovin.shopkeepers.text.Text;
+import com.nisovin.shopkeepers.util.EntityUtils;
+import com.nisovin.shopkeepers.util.TextUtils;
+
+/**
+ * Produces lots of damage events (with a damage of <code>0</code>) in quick succession.
+ * <p>
+ * This can for example be used to test the performance of the involved event handlers.
+ * <p>
+ * This may conflict with anti-cheat detection plugins.
+ */
+class CommandTestDamage extends PlayerCommand {
+
+	private static final String ARGUMENT_DAMAGE = "damage";
+	private static final String ARGUMENT_TIMES_PER_TICK = "times-per-tick";
+	private static final String ARGUMENT_DURATION_TICKS = "duration-ticks";
+
+	private final SKShopkeepersPlugin plugin;
+
+	CommandTestDamage(SKShopkeepersPlugin plugin) {
+		super("testDamage");
+		this.plugin = plugin;
+
+		// Set permission:
+		this.setPermission(ShopkeepersPlugin.DEBUG_PERMISSION);
+
+		// Set description:
+		this.setDescription(Text.of("Produces damage events for the targetted entity."));
+
+		// Hidden debugging command:
+		this.setHiddenInParentHelp(true);
+
+		// Arguments:
+		this.addArgument(new DefaultValueFallback<>(new BoundedDoubleArgument(ARGUMENT_DAMAGE, 0.0D, Double.MAX_VALUE), 0.0D));
+		this.addArgument(new DefaultValueFallback<>(new PositiveIntegerArgument(ARGUMENT_TIMES_PER_TICK), 1));
+		this.addArgument(new DefaultValueFallback<>(new PositiveIntegerArgument(ARGUMENT_DURATION_TICKS), 20 * 10));
+	}
+
+	@Override
+	protected void execute(CommandInput input, CommandContextView context) throws CommandException {
+		Player player = (Player) input.getSender();
+		double damage = context.get(ARGUMENT_DAMAGE);
+		assert damage >= 0.0D;
+		int timesPerTick = context.get(ARGUMENT_TIMES_PER_TICK);
+		assert timesPerTick > 1;
+		int durationTicks = context.get(ARGUMENT_DURATION_TICKS);
+		assert durationTicks > 1;
+
+		LivingEntity target = (LivingEntity) EntityUtils.getTargetedEntity(player, (entity) -> entity instanceof LivingEntity);
+		if (target == null) {
+			player.sendMessage(ChatColor.RED + "No living entity targetted!");
+			return;
+		}
+
+		// Start damage task:
+		player.sendMessage(TextUtils.colorize("&aStarting damage task: Target: &e" + target.getName() + "&a, Damage: &e"
+				+ TextUtils.DECIMAL_FORMAT.format(damage) + "&a, Per tick: &e" + timesPerTick + "&a, Duration &e"
+				+ durationTicks + " ticks &a..."));
+
+		new BukkitRunnable() {
+
+			private int tickCounter = 0;
+
+			@Override
+			public void run() {
+				boolean playerValid = player.isValid();
+				if (tickCounter >= durationTicks || !playerValid || !target.isValid()) {
+					// We are done:
+					if (playerValid) {
+						player.sendMessage(ChatColor.GREEN + "... Done");
+					}
+					this.cancel();
+					return;
+				}
+
+				// Apply damage:
+				for (int i = 0; i < timesPerTick; ++i) {
+					// Reset damage cooldown:
+					target.setNoDamageTicks(0);
+					target.setLastDamage(0.0D);
+
+					// Damage:
+					target.damage(damage, player);
+
+					// Abort if the entity died:
+					if (target.isDead()) {
+						break;
+					}
+				}
+
+				tickCounter += 1;
+
+				// Periodic progress feedback:
+				if ((tickCounter % 20) == 0) {
+					player.sendMessage(ChatColor.GRAY + "... (" + ChatColor.YELLOW + tickCounter + ChatColor.GRAY + " / "
+							+ ChatColor.YELLOW + durationTicks + ChatColor.GRAY + ")");
+				}
+			}
+		}.runTaskTimer(plugin, 1L, 1L);
+	}
+}